backend/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
fastapi
uvicorn
python-dotenv
pydantic
starlette
itsdangerous